+++ src/libs/test/AxelarQueryAPI.spec.ts
@@ -49,6 +49,18 @@ describe("AxelarQueryAPI", () => {
       expect(response.fee?.denom).toEqual(assetDenom);
       expect(response.fee?.amount).toBeDefined();
     });
+
+    test("it should suggest a chain id when passing chain name", async () => {
+      const [sourceChainName, destinationChainName, assetDenom, amount] = [
+        "osmosis",
+        "polygon",
+        "uusd",
+        100000000,
+      ];
+      expect(
+        api.getTransferFee(sourceChainName, destinationChainName, assetDenom, amount)
+      ).rejects.toThrow("Invalid chain identifier for osmosis. Did you mean osmosis-4");
+    });
   });
 
   describe("getGasPrice", () => {
@@ -146,14 +158,14 @@ describe("AxelarQueryAPI", () => {
   describe("throwIfInactiveChain", () => {
     test("It should throw if the chain does not get included in a active-chains list", async () => {
       jest.spyOn(api, "getActiveChains").mockResolvedValue(["avalanche", "polygon"]);
-      await expect(api.throwIfInactiveChain("ethereum")).rejects.toThrowError(
+      await expect(api.throwIfInactiveChains(["ethereum"])).rejects.toThrowError(
         "Chain ethereum is not active"
       );
     });
 
     test("It should throw if the chain does not get included in a active-chains list", async () => {
       jest.spyOn(api, "getActiveChains").mockResolvedValue(["avalanche", "polygon"]);
-      await expect(api.throwIfInactiveChain("avalanche")).resolves.toBeUndefined();
+      await expect(api.throwIfInactiveChains(["avalanche"])).resolves.toBeUndefined();
     });
   });
 });